+++ example/cli/src/app/fulltree/fulltree.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input } from '@angular/core';
-import { TreeNode, TREE_ACTIONS, KEYS, IActionMapping, ITreeOptions } from 'angular-tree-component';
+import { TreeNode, TreeModel, TREE_ACTIONS, KEYS, IActionMapping, ITreeOptions } from 'angular-tree-component';
 
 const actionMapping:IActionMapping = {
   mouse: {
@@ -45,7 +45,8 @@ const actionMapping:IActionMapping = {
       [nodes]="nodes"
       [options]="customTemplateStringOptions"
       [focused]="true"
-      (onEvent)="onEvent($event)"
+      (event)="onEvent($event)"
+      (initialized)="onInitialized(tree)"
       
     >
       <ng-template #treeNodeTemplate let-node>
@@ -136,6 +137,7 @@ export class FullTreeComponent {
             {
               name: 'child2.1',
               subTitle: 'new and improved',
+              uuid: '11',
               hasChildren: false
             }, {
               name: 'child2.2',
@@ -220,7 +222,14 @@ export class FullTreeComponent {
     getChildren: this.getChildren.bind(this),
     actionMapping,
     nodeHeight: 23,
-    allowDrag: true,
+    allowDrag: (node) => {
+      // console.log('allowDrag?');
+      return true;
+    },
+    allowDrop: (node) => {
+      // console.log('allowDrop?');
+      return true;
+    },
     useVirtualScroll: true,
     animateExpand: true,
     animateSpeed: 30,
@@ -228,6 +237,10 @@ export class FullTreeComponent {
   }
   onEvent(event) {
     console.log(event);
+  }
+
+  onInitialized(tree) {
+    // tree.treeModel.getNodeById('11').setActiveAndVisible();
   }
 
   go($event) {